chore: extract the adapt-runtime-plan part of the algorithm out of Engine
It will have to be significantly modified for the lazy-deploy mode, so best have
the current known-to-work version in a cleanly separated class.
