+++ lib/syskit/network_generation.rb
@@ -17,5 +17,6 @@ require "syskit/network_generation/engine"
 require "syskit/network_generation/logger"
 require "syskit/network_generation/merge_solver"
 require "syskit/network_generation/resolution_error_handler"
+require "syskit/network_generation/runtime_network_adaptation"
 require "syskit/network_generation/system_network_deployer"
 require "syskit/network_generation/system_network_generator"
